mobile/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
email-validator==2.1.0
httpx==0.27.0
slowapi==0.1.9
websockets==12.0
